Rename generate_wheel to manage_dependencies
